+++ sites/x6-sites/src/tutorial/getting-started/react-shape/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Graph, Node } from '@antv/x6'
 import { register } from '@antv/x6-react-shape'
 import { Dropdown } from 'antd'
-import styles from './index.less'
+import './index.less'
 
 const CustomComponent = ({ node }: { node: Node }) => {
   const label = node.prop('label')
@@ -26,7 +26,7 @@ const CustomComponent = ({ node }: { node: Node }) => {
       }}
       trigger={['contextMenu']}
     >
-      <div className={styles['custom-react-node']}>{label}</div>
+      <div className="custom-react-node">{label}</div>
     </Dropdown>
   )
 }
@@ -92,8 +92,8 @@ export default class Example extends React.Component {
 
   render() {
     return (
-      <div className={styles.app}>
-        <div className={styles['app-content']} ref={this.refContainer} />
+      <div className="react-shape-app">
+        <div className="app-content" ref={this.refContainer} />
       </div>
     )
   }