+++ sites/x6-sites/src/tutorial/intermediate/group/collapsable/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Graph } from '@antv/x6'
 import { Group } from './shape'
-import styles from './index.less'
+import './index.less'
 
 Graph.registerNode(
   'custom-group-node',
@@ -146,8 +146,8 @@ export default class Example extends React.Component {
 
   render() {
     return (
-      <div className={styles.app}>
-        <div className={styles['app-content']} ref={this.refContainer} />
+      <div className="collapsable-app">
+        <div className="app-content" ref={this.refContainer} />
       </div>
     )
   }