+++ sites/x6-sites/src/tutorial/intermediate/react/update/index.tsx
@@ -2,12 +2,12 @@ import React from 'react'
 import { Graph, Node } from '@antv/x6'
 import { register } from '@antv/x6-react-shape'
 import { Progress } from 'antd'
-import styles from './index.less'
+import './index.less'
 
 const NodeComponent = ({ node }: { node: Node }) => {
   const { progress } = node.getData()
   return (
-    <div className={styles['react-node']}>
+    <div className="react-node">
       <Progress type="circle" percent={progress} width={80} />
     </div>
   )
@@ -57,8 +57,8 @@ export default class Example extends React.Component {
 
   render() {
     return (
-      <div className={styles.app}>
-        <div className={styles['app-content']} ref={this.refContainer} />
+      <div className="react-update-app ">
+        <div className="app-content" ref={this.refContainer} />
       </div>
     )
   }