+++ sites/x6-sites/src/tutorial/intermediate/tools/basic/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Graph } from '@antv/x6'
-import styles from './index.less'
+import './index.less'
 
 Graph.registerNode(
   'custom-tools-node',
@@ -70,8 +70,8 @@ export default class Example extends React.Component {
 
   render() {
     return (
-      <div className={styles.app}>
-        <div className={styles['app-content']} ref={this.refContainer} />
+      <div className="tools-basic-app">
+        <div className="app-content" ref={this.refContainer} />
       </div>
     )
   }